컴포넌트 간소화
